# src/main/api/openapi.yaml
openapi: 3.0.3
info:
  title: Lyra 企业级文档管理系统 API
  description: |
    Lyra是一个企业级云原生文档管理系统，提供带有版本控制的文档管理服务。
    
    ## 核心功能
    - 用户认证与授权管理
    - 文件上传、下载、管理
    - 文件夹结构化管理
    - 文件搜索与版本控制
    - WebDAV协议支持
    - 权限管理与空间隔离
    
    ## 技术架构
    - **后端**: Spring Boot 3.5.x + Spring Security + JPA
    - **数据库**: SQLite/MySQL/PostgreSQL + Redis
    - **部署**: Docker + Kubernetes + Helm
    
    ## API特性
    - RESTful设计规范
    - JWT令牌认证
    - 统一响应格式
    - 完整的错误处理
    - 分页和排序支持
    - 文件流传输优化
    
  version: "1.0.0"
  termsOfService: https://github.com/your-org/lyra/blob/main/TERMS.md
  contact:
    name: Lyra API Support
    url: https://github.com/your-org/lyra
    email: support@lyra.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://api-dev.lyra.example.com
    description: 开发环境
  - url: https://api-staging.lyra.example.com  
    description: 预发布环境
  - url: https://api.lyra.example.com
    description: 生产环境

# ==================== 安全方案 ====================
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT访问令牌认证。
        
        **获取令牌**: 通过 `/api/auth/login` 端点登录获取访问令牌。
        
        **使用方式**: 在请求头中添加 `Authorization: Bearer <token>`
        
        **令牌刷新**: 使用 `/api/auth/refresh` 端点刷新过期令牌。

  # ==================== 通用响应格式 ====================
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 操作是否成功
          example: true
        message:
          type: string
          description: 响应消息
          example: "操作成功"
        data:
          description: 响应数据，根据具体接口而定
        timestamp:
          type: integer
          format: int64
          description: 时间戳(毫秒)
          example: 1642780800000
        errors:
          type: array
          description: 错误信息列表(仅在失败时出现)
          items:
            type: string
          example: ["参数验证失败", "用户名不能为空"]
      required:
        - success
        - message

    # 分页响应格式
...
    LoginRequest:
      type: object
      required:
        - usernameOrEmail
        - password
      properties:
        usernameOrEmail:
          type: string
          description: 用户名或邮箱
          example: "admin@lyra.com"
        password:
          type: string
          format: password
          description: 密码
          example: "password123"
        rememberMe:
          type: boolean
          description: 是否记住登录状态
          default: false

    # 登录响应
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: 访问令牌
          example: "eyJhbGciOiJIUzI1NiIs..."
        refreshToken:
          type: string
          description: 刷新令牌
          example: "eyJhbGciOiJIUzI1NiIs..."
        tokenType:
          type: string
          description: 令牌类型
          example: "Bearer"
        expiresIn:
          type: integer
          description: 令牌过期时间(秒)
          example: 3600
        user:
          $ref: '#/components/schemas/UserInfo'

    # 注册请求
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          description: 用户名
          pattern: '^[a-zA-Z0-9_]{3,20}$'
          example: "newuser"
        email:
          type: string
          format: email
          description: 邮箱地址
          example: "newuser@example.com"
        password:
          type: string
          format: password
          minLength: 8
          description: 密码(至少8位)
          example: "password123"
        displayName:
          type: string
          description: 显示名称
          example: "新用户"

    # 用户信息
    UserInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 用户ID
          example: 1
        username:
          type: string
          description: 用户名
          example: "admin"
        email:
          type: string
          format: email
          description: 邮箱
          example: "admin@lyra.com"
        displayName:
          type: string
          description: 显示名称
          example: "系统管理员"
        avatar:
          type: string
          format: uri
          description: 头像URL
          example: "https://example.com/avatar.jpg"
        roles:
          type: array
          description: 用户角色
          items:
            type: string
          example: ["ADMIN", "USER"]
        status:
          type: string
          enum: [ACTIVE, DISABLED, LOCKED]
          description: 账户状态
          example: "ACTIVE"
        createdAt:
          type: string
          format: date-time
          description: 创建时间
          example: "2024-01-01T12:00:00Z"

    # 刷新令牌请求
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: 刷新令牌
          example: "eyJhbGciOiJIUzI1NiIs..."

    # ==================== 文件相关数据模型 ====================
    
    # 文件信息
...
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [false]
            errors:
              type: array
              items:
                type: string

  # ==================== 通用参数 ====================
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "请求参数验证失败"
            errors: ["用户名不能为空"]
            timestamp: 1642780800000
    Unauthorized:
      description: 未授权访问
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "访问令牌无效或已过期"
            timestamp: 1642780800000
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - 认证管理
      summary: 用户登录
      description: |
        用户通过用户名/邮箱和密码进行登录认证。
        
        **注意事项**:
        - 支持用户名或邮箱登录
        - 密码错误次数过多将导致账户锁定
        - 返回的访问令牌有效期为1小时
        - 刷新令牌有效期为7天
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              usernameOrEmail: "admin@lyra.com"
              password: "password123"
              rememberMe: true
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
              example:
                success: true
                message: "登录成功"
                data:
                  accessToken: "eyJhbGciOiJIUzI1NiIs..."
                  refreshToken: "eyJhbGciOiJIUzI1NiIs..."
                  tokenType: "Bearer"
                  expiresIn: 3600
                  user:
                    id: 1
                    username: "admin"
                    email: "admin@lyra.com"
                    displayName: "系统管理员"
                    roles: ["ADMIN", "USER"]
                    status: "ACTIVE"
                timestamp: 1642780800000
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: 登录失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "用户名或密码错误"
                timestamp: 1642780800000
  /api/auth/register:
    post:
      tags:
        - 认证管理
      summary: 用户注册
      description: |
        新用户注册账户。
        
        **注意事项**:
        - 用户名必须是3-20位字母数字下划线
        - 密码至少8位
        - 邮箱地址必须唯一
        - 注册后需要邮箱验证(如果启用)
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: "newuser"
              email: "newuser@example.com"
              password: "password123"
              displayName: "新用户"
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: 用户名或邮箱已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/refresh:
    post:
      tags:
        - 认证管理
      summary: 刷新访问令牌
      description: |
        使用刷新令牌获取新的访问令牌。
        
        **注意事项**:
        - 当访问令牌过期时使用
        - 刷新令牌只能使用一次
        - 返回新的访问令牌和刷新令牌
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 令牌刷新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
                          tokenType:
                            type: string
                          expiresIn:
                            type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/logout:
    post:
      tags:
        - 认证管理
      summary: 用户登出
      description: |
        用户登出，令牌加入黑名单。
        
        **注意事项**:
        - 当前访问令牌将被加入黑名单
        - 登出后需要重新登录获取新令牌
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "登出成功"
                timestamp: 1642780800000

  # ==================== 文件管理API ====================
...